+++ pecha_api/app.py
@@ -6,7 +6,6 @@ from pecha_api.auth.auth_service import retrieve_client_info
 
 from pecha_api.db.mongo_database import lifespan
 from pecha_api.auth import auth_views
-from pecha_api.sheets import sheets_views
 from pecha_api.collections import collections_views
 from pecha_api.terms import terms_views
 from pecha_api.texts import texts_views
@@ -51,7 +50,6 @@ api = FastAPI(
     lifespan=lifespan
 )
 api.include_router(auth_views.auth_router)
-api.include_router(sheets_views.sheets_router)
 api.include_router(collections_views.collections_router)
 api.include_router(terms_views.terms_router)
 api.include_router(texts_views.text_router)
