--- pecha_api/sheets/sheets_enum.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-class SortOrder(Enum):
-    ASC = "asc"
-    DESC = "desc"
-
-class SortBy(Enum):
-    CREATED_DATE = "created_date"
-    PUBLISHED_DATE = "published_date"
