+++ pecha_api/texts/segments/segments_response_models.py
@@ -7,19 +7,6 @@ from .segments_enum import SegmentType
 from pecha_api.texts.texts_response_models import TextDTO
 
 
-class CreateSegment(BaseModel):
-    pecha_segment_id: Optional[str] = None
-    content: str
-    pecha_segment_id: Optional[str] = None
-    type: SegmentType
-    mapping: Optional[List[Mapping]] = []
-
-
-class CreateSegmentRequest(BaseModel):
-    text_id: str
-    segments: List[CreateSegment]
-
-
 class MappingResponse(BaseModel):
     text_id: str
     segments: List[str]
