+++ pecha_api/texts/segments/segments_service.py
@@ -1,23 +1,11 @@
-from pecha_api.error_contants import ErrorConstants
 from .segments_repository import (
-    create_segment,
     get_segments_by_ids,
-    get_segments_by_text_id,
-    delete_segments_by_text_id,
 )
 from .segments_response_models import (
-    CreateSegmentRequest,
-    SegmentResponse,
-    MappingResponse,
     SegmentDTO,
 )
 
 from pecha_api.cache.cache_enums import CacheType
-
-from fastapi import HTTPException
-from starlette import status
-
-from ..texts_utils import TextUtils
 
 from typing import List, Dict
 
@@ -25,8 +13,6 @@ from .segments_cache_service import (
     get_segments_details_by_ids_cache,
     set_segments_details_by_ids_cache,
 )
-
-from ...users.users_service import validate_user_exists
 
 async def get_segments_details_by_ids(segment_ids: List[str]) -> Dict[str, SegmentDTO]:
     cached_data: Dict[str, SegmentDTO] = await get_segments_details_by_ids_cache(segment_ids=segment_ids, cache_type=CacheType.SEGMENTS_DETAILS)
@@ -38,30 +24,3 @@ async def get_segments_details_by_ids(segment_ids: List[str]) -> Dict[str, Segme
     await set_segments_details_by_ids_cache(segment_ids=segment_ids, cache_type=CacheType.SEGMENTS_DETAILS, data=segments)
     
     return segments
-
-
-async def create_new_segment(create_segment_request: CreateSegmentRequest, token: str) -> SegmentResponse:
-    is_valid_user = validate_user_exists(token=token)
-    if is_valid_user:
-        await TextUtils.validate_text_exists(text_id=create_segment_request.text_id)
-        new_segment = await create_segment(create_segment_request=create_segment_request)
-        segments =  [
-            SegmentDTO(
-                id=str(segment.id),
-                pecha_segment_id=str(segment.pecha_segment_id),
-                text_id=segment.text_id,
-                content=segment.content,
-                mapping= [MappingResponse(**mapping.model_dump()) for mapping in segment.mapping],
-                type=segment.type
-            )
-            for segment in new_segment
-        ]
-        return SegmentResponse(segments=segments)
-    else:
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail=ErrorConstants.TOKEN_ERROR_MESSAGE)
-
-async def remove_segments_by_text_id(text_id: str):
-    is_valid_text = await TextUtils.validate_text_exists(text_id=text_id)
-    if not is_valid_text:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=ErrorConstants.TEXT_NOT_FOUND_MESSAGE)
-    return await delete_segments_by_text_id(text_id=text_id)
