+++ pecha_api/texts/texts_enums.py
@@ -12,6 +12,16 @@ class TextType(Enum):
 class TextTypes(Enum):
     VERSIONS = "versions"
 
+class SortOrder(Enum):
+    ASC = "asc"
+    DESC = "desc"
+
+
+class SortBy(Enum):
+    CREATED_DATE = "created_date"
+    PUBLISHED_DATE = "published_date"
+
+
 class PaginationDirection(Enum):
     NEXT = "next"
     PREVIOUS = "previous"
