+++ pecha_api/texts/texts_models.py
@@ -7,11 +7,6 @@ from .texts_response_models import Section
 from pydantic import Field
 from beanie import Document
 
-from pecha_api.sheets.sheets_enum import (
-    SortBy, 
-    SortOrder
-)
-
 from .texts_enums import TextType
 from .texts_response_models import TextDTO, TableOfContentType
 
@@ -211,56 +206,3 @@ class Text(Document):
         return await cls.find_one(cls.id == text_id).delete()
 
 
-    @classmethod
-    async def get_sheets(
-        cls, 
-        published_by: Optional[str] = None,
-        is_published: Optional[bool] = None,
-        sort_by: Optional[SortBy] = None,
-        sort_order: Optional[SortOrder] = None,
-        skip: int = 0, 
-        limit: int = 10
-    ) -> List["Text"]:
-        query = {"type": TextType.SHEET}
-            
-        if published_by is not None:
-            query["published_by"] = published_by
-            
-        if is_published is not None:
-            query["is_published"] = is_published
-
-        mongo_query = cls.find(query)
-        if sort_by:
-            field = {
-                SortBy.CREATED_DATE: "created_date",
-                SortBy.PUBLISHED_DATE: "published_date",
-            }
-            sort_field = field.get(sort_by)
-            if sort_field:
-                if sort_order == SortOrder.DESC:
-                    sort_string = f"-{sort_field}"
-                else:
-                    sort_string = sort_field
-            
-                mongo_query = mongo_query.sort(sort_string)
-
-        mongo_query = mongo_query.skip(skip).limit(limit)
-
-
-        texts = await mongo_query.to_list()
-
-        return texts
-
-
-    @classmethod
-    async def get_published_sheets_count_from_db(
-        cls,
-        email: Optional[str] = None
-    ) -> int:
-        query = {
-            "type": TextType.SHEET, 
-            "is_published": True
-        }
-        if email is not None:
-            query["published_by"] = email
-        return await cls.find(query).count()
